requirements.txt
Flask
datetime
uuid
Flask-SQLAlchemy
PyJWT
psycopg2
flask-bcrypt
marshmallow
gunicorn